+++ packages/web/src/errors/errors-list.ts
@@ -1,3 +1,5 @@
+import { title } from "process";
+
 export const ERROR_PREFIX = "ABLDR";
 
 export interface ErrorDescriptor {
@@ -193,6 +195,13 @@ Please check algob output for more details.`,
 		message: `Application(name = %appName%) not found in checkpoint`,
 		title: "App by name not found in checkpoint",
 		description: "App by name not found in checkpoint",
+	},
+	TRANSACTION_LENGTH_ERROR: {
+		number: 21,
+		message: "Invalid length of transaction parameter",
+		title: "Transaction parameter length",
+		description:
+			"Transaction parameter needs to have between 1 and 16 transactions, got: %length%",
 	},
 };