+++ packages/web/src/lib/web-mode.ts
@@ -1,9 +1,18 @@
 import algosdk, { SuggestedParams, Transaction } from "algosdk";
 
 import { AlgoSigner, JsonPayload, WalletTransaction } from "../algo-signer-types";
-import { ExecParams, Sign, SignType, TxParams } from "../types";
+import { BuilderError, ERRORS } from "../errors/errors";
+import {
+	ExecParams,
+	isSDKTransactionAndSign,
+	Sign,
+	SignType,
+	TransactionAndSign,
+	TxParams,
+} from "../types";
+import { WAIT_ROUNDS } from "./constants";
 import { log } from "./logger";
-import { getFromAddress, mkTransaction } from "./txn";
+import { mkTransaction } from "./txn";
 
 const CONFIRMED_ROUND = "confirmed-round";
 const LAST_ROUND = "last-round";
@@ -21,15 +30,18 @@ export class WebMode {
 	 * wait for confirmation for transaction using transaction id
 	 * @param txId Transaction id
 	 */
-	async waitForConfirmation(txId: string): Promise<JsonPayload> {
+	async waitForConfirmation(
+		txId: string
+	): Promise<algosdk.modelsv2.PendingTransactionResponse> {
 		const response = await this.algoSigner.algod({
 			ledger: this.chainName,
 			path: "/v2/status",
 		});
 		log(response);
-		let lastround = response[LAST_ROUND] as number;
+		const startRound = response[LAST_ROUND] as number;
+		let currentRound = startRound;
 		// eslint-disable-next-line no-constant-condition
-		while (true) {
+		while (currentRound < startRound + WAIT_ROUNDS) {
 			const pendingInfo = await this.algoSigner.algod({
 				ledger: this.chainName,
 				path: `/v2/transactions/pending/${txId}`,
@@ -38,15 +50,16 @@ export class WebMode {
 				pendingInfo[CONFIRMED_ROUND] !== null &&
 				(pendingInfo[CONFIRMED_ROUND] as number) > 0
 			) {
-				return pendingInfo;
+				return pendingInfo as unknown as algosdk.modelsv2.PendingTransactionResponse;
 			}
 			// TODO: maybe we should use "sleep" instead of pinging a node again?
-			lastround++;
+			currentRound += 1;
 			await this.algoSigner.algod({
 				ledger: this.chainName,
-				path: `/v2/status/wait-for-block-after/${lastround}`, // eslint-disable-line @typescript-eslint/restrict-template-expressions
+				path: `/v2/status/wait-for-block-after/${currentRound}`, // eslint-disable-line @typescript-eslint/restrict-template-expressions
 			});
 		}
+		throw new Error(`Transaction not confirmed after ${WAIT_ROUNDS} rounds`);
 	}
 
 	/**
@@ -129,14 +142,23 @@ export class WebMode {
 
 	/**
 	 * Execute single transaction or group of transactions (atomic transaction)
-	 * @param execParams transaction parameters or atomic transaction parameters
+	 * @param transactions transaction parameters, atomic transaction parameters
+	 * or TransactionAndSign object(SDK transaction object and signer parameters)
 	 */
-	async executeTx(execParams: ExecParams[]): Promise<JsonPayload> {
+	async executeTx(
+		transactions: ExecParams[] | TransactionAndSign[]
+	): Promise<algosdk.modelsv2.PendingTransactionResponse> {
 		let txns: Transaction[] = [];
-		if (execParams.length > 16) {
-			throw new Error("Maximum size of an atomic transfer group is 16");
+		if (transactions.length > 16 || transactions.length == 0) {
+			throw new BuilderError(ERRORS.GENERAL.TRANSACTION_LENGTH_ERROR, {
+				length: transactions.length,
+			});
 		}
 
+		if (isSDKTransactionAndSign(transactions[0]))
+			throw new Error("We don't support this case now");
+
+		const execParams = transactions as ExecParams[];
 		for (const [_, txn] of execParams.entries()) {
 			txns.push(mkTransaction(txn, await this.getSuggestedParams(txn.payFlags)));
 		}
@@ -183,7 +205,7 @@ export class WebMode {
 
 	/** @deprecated */
 	async executeTransaction(execParams: ExecParams | ExecParams[]): Promise<JsonPayload> {
-		if (Array.isArray(execParams)) return this.executeTx(execParams);
-		else return this.executeTx([execParams]);
+		if (Array.isArray(execParams)) return this.executeTx(execParams) as unknown as JsonPayload;
+		else return this.executeTx([execParams]) as unknown as JsonPayload;
 	}
 }